Fix pause after goodbye message before program exit
When provided a bye command, the application is expected to display the
bot's goodbye message for three seconds prior to closing the window. The
application pauses and exits without showing the goodbye message.

This occurs due to the use of Thread.sleep to create a pause in
execution, which puts the thread to sleep before the UI updates, and the
call to System.exit immediately after. Thus, a different wait mechanism
is necessary.

Let's use the PauseTransition within JavaFX, which allows us to pause
the program without blocking the UI thread.

Let's also use Platform.exit to close the application instead, as it
will guarantee that in the future, any UI updates can be completed
before closing.