Add assertions to storage package to avoid invalid states
The storage package allows users to instantiate an instance of
storage.Storage with a null type, empty filepath, and even save a null
object to a file.

This can lead to invalid states, since type information is required for
typecasting when loading from a file. Furthermore, saving a null object
should not be allowed as the user should be forced to delete the file
instead. Lastly, an empty filepath should not be allowed as it does not
provide a valid file to save to.

Let's make use of assertions to enforce the aforementioned constraints.

We choose to use assertions rather than exceptions for these cases, as
these are invalid states that the implementer should not allow to
happen, rather than exceptional errors to be handled.

Signed-off-by: Reuben Thomas <reubenthomas123@gmail.com>