Replace public constructors with factory methods
The codebase makes use of public extensively, which is often regarded as
bad practice and less maintainable, since the code associated with the
construction of an object, and the code used to interface with the
object are the same.

Thus, if the constructor needs to be changed in the future, as is
likely, all instances of its usage in other files must also be changed,
which makes it less maintainable.

Let's implement factory methods associated with the creation of objects,
that can be called from other classes. Furthermore, let's make the
constructors private so that they may not accidentally be used to
instantiate objects.

Factory methods are chosen instead of the builder pattern, since it
allows us to implement it quicker, and for less additional code.

Signed-off-by: Reuben Thomas <reubenthomas123@gmail.com>