Assert non-empty response message in CommandResult
The CommandResult class does not enforce on instantiation, whether the
response or error optional strings should be empty, and if present, that
they are not blank.

Let's add assertions to the constructor to enforce that for results
associated with invalid usage, and unexpected errors must have a
non-empty error message and an empty response message. Likewise, for a
result associated with successful execution, the response must include a
non-empty response message and an empty error message.

We choose to enforce this through assertions rather than throwing an
exception, since this is a case of incorrect functioning or modification
of the class.

In the future, a factory method may be used to further guarantee as the
only exposed methods to instantiate a CommandResult, while the
constructor continues to manage the current assertions implemented.