Add assertions to enforce unique command names
A user should not be able to add multiple commands with the same name.
The CommandHandler class method to add a command simply replaces the
existing command with the new one without any checks or indication to
the user.

Let's make use of assertions to enforce that the new command being added
is unique, and does not already exist in the command map.

We make use of assertions instead of exceptions in this case, since it
is a case of incorrect usage of the class by a programmer when defining
a set of commands associated with a command handler prior to execution.
It is critical to prevent existing commands from being mistakenly
overwritten.